+++ packages/fork-tests/src/index.ts
@@ -1,4 +1,3 @@
-export * from './forkTests';
 export * from './network';
 export * from './signers';
 export * from './test';
