--- packages/fork-tests/tests/20221116-gearbox-rebalanced-linear-pool/readme.md
@@ -1,8 +1,0 @@
-# 2022-11-16 - Gearbox Rebalanced Linear Pool
-
-Deployment of the `GearboxLinearPoolFactory`, for Linear Pools with a Gearbox Diesel Token (dToken).
-
-## Useful Files
-- [`GearboxLinearPool` ABI](./abi/GearboxLinearPool.json)
-- [`GearboxLinearPoolFactory` ABI](./abi/GearboxLinearPoolFactory.json)
-- [`GearboxLinearPoolRebalancer` ABI](./abi/GearboxLinearPoolRebalancer.json)
