+++ packages/fork-tests/tests/20230102-gearbox-rebalanced-linear-pool/input.ts
@@ -1,0 +1,31 @@
+import Task, { TaskMode } from '../../src/task';
+import { MONTH } from '@orbcollective/shared-dependencies/time';
+
+export type GearboxLinearPoolDeployment = {
+  Vault: string;
+  BalancerQueries: string;
+  ProtocolFeePercentagesProvider: string;
+  WETH: string;
+  FactoryVersion: string;
+  PoolVersion: string;
+  InitialPauseWindowDuration: number;
+  BufferPeriodDuration: number;
+};
+
+const Vault = new Task('20210418-vault', TaskMode.READ_ONLY);
+const BalancerQueries = new Task('20220721-balancer-queries', TaskMode.READ_ONLY);
+const ProtocolFeePercentagesProvider = new Task('20220725-protocol-fee-percentages-provider', TaskMode.READ_ONLY);
+const WETH = new Task('00000000-tokens', TaskMode.READ_ONLY);
+
+const BaseVersion = { version: 1, deployment: '20230102-gearbox-rebalanced-linear-pool' };
+
+export default {
+  Vault,
+  BalancerQueries,
+  ProtocolFeePercentagesProvider,
+  WETH,
+  FactoryVersion: JSON.stringify({ name: 'GearboxLinearPoolFactory', ...BaseVersion }),
+  PoolVersion: JSON.stringify({ name: 'GearboxLinearPool', ...BaseVersion }),
+  InitialPauseWindowDuration: MONTH * 3,
+  BufferPeriodDuration: MONTH,
+};
