+++ packages/fork-tests/tests/20230102-gearbox-rebalanced-linear-pool/output/test.json
@@ -1,0 +1,4 @@
+{
+  "AaveLinearPoolFactory": "0x7C8BaafA542c57fF9B2B90612bf8aB9E86e22C09",
+  "GearboxLinearPoolFactory": "0x90c84237fDdf091b1E63f369AF122EB46000bc70"
+}
