+++ packages/fork-tests/tests/20230102-gearbox-rebalanced-linear-pool/readme.md
@@ -1,0 +1,9 @@
+# 2022-12-07 - Gearbox Rebalanced Linear Pool
+
+Deployment of the `GearboxLinearPoolFactory`, for Linear Pools with a wrapped diesel token.
+
+## Useful Files
+
+- [`GearboxLinearPool` artifact](./artifact/GearboxLinearPool.json)
+- [`GearboxLinearPoolFactory` artifact](./artifact/GearboxLinearPoolFactory.json)
+- [`GearboxLinearPoolRebalancer` artifact](./artifact/GearboxLinearPoolRebalancer.json)
